requirements.txt
gradio~=5.22.0
moviepy~=1.0.3
requests~=2.32.3
python-dotenv~=1.0.1